+++ src/Typin/Typin.Core/Attributes/CommandParameterAttribute.cs
@@ -1,9 +1,6 @@
 ﻿namespace Typin.Attributes
 {
     using System;
-    using Typin.Binding;
-    using Typin.Internal.Exceptions;
-    using Typin.Internal.Extensions;
 
     /// <summary>
     /// Annotates a property that defines a command parameter.